+++ HISTORY.md
@@ -1,5 +1,10 @@
 History of features added to LiChess Tools in time (reversed order)
 
+13 May
+
+- More Calculation Trainer features
+- chess.com shuiq board
+
 05 May
 
 - Calculation trainer